Very Large Update
Created a Master Controller to control the whole game.

Created a Combat Class Controller to control the Game Logic related to Combat.

Created a Listeners Controller to listen to all buttons in the game.

Imported the Handlebars class to start using Templates.

Had to make a minor change to the generateEnemy Function.

Made a change to the levelup enemy function.

Moved isDead from a private method to a public method.

Updated the Salamanders levelup guide.

Updated the Zombie class Constructor.

Updated the Skeleton class Constructor.

Made some minor changes to the CharacterClass.

Added some CSS.